Fixup GetTeamId().
Previously returned raw Teams enum class (not sure how it compiled in the first place). Added explicit cast to uint8 to correct.

+++ Source/RushReborn/Public/Towers/Tower.cpp
@@ -3,6 +3,6 @@
 
 uint8 ATower::GetTeamId()
 {
-	return ETeams::Player;
+	return (uint8)ETeams::Player;
 }
